+++ SharedTrips/Services/Trips/TripServiceListingModel.cs
@@ -25,7 +25,7 @@ namespace SharedTrips.Services.Trips
 
         public string DriverPictureUrl { get; set; }
 
-        public int DriverRating { get; set; }
+        public double DriverRating { get; set; }
 
         public int CarId { get; set; }
     }
